+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to `knotx-fragments` will be documented in this file.
 ## Unreleased
 List of changes that are finished but not yet released in any final version.
 - [PR-83](https://github.com/Knotx/knotx-fragments/pull/83) - Adding node log to inline body action.
+- [PR-82](https://github.com/Knotx/knotx-fragments/pull/82) - Adding logging to in-memory cache action.
 - [PR-60](https://github.com/Knotx/knotx-fragments/pull/60) - Adding node log to circuit breaker action. Enforce fallback on error.
+
 
 ## Version 2.1.0
 - [PR-62](https://github.com/Knotx/knotx-fragments/pull/62) - It introduces task & graph node factories. It is connected with issue #49.